refactor: getStateValueString -> stepIdConverter
